Chicago sales tax revenue by ZIP code for 2020
Provided by the Illinois Department of Revenue 

My REQUEST:
````
A copy of all data, of which is in aggregate form, that has been used or shared either with, or by, your agency that includes the retail sales tax information of one or more Illinois businesses, which has a presence in the City of Chicago, aggregated at the Census tract level. Please limit the scope of this request to retail sales tax collected in 2020. 

Per Retailers' Occupation Tax Act, 35 ILCS 120/11 Sec. 11, "reasonable statistics" can be made available for public records- ie, aggregated retails tax information. The release of this information can be done without the violation of reciprocal agreements of information similar to [1] "Nothing in this Act prevents the Director of Revenue from publishing or making available to the public the names and addresses of persons filing returns under this Act, or reasonable statistics concerning the operation of the tax by grouping the contents of returns so the information in any individual return is not disclosed."

I understand that this request may be a bit broad - I can be reached at 312-945-9621 to discuss the clarification and limiting of scope for this request.
If there are no records responsive to the request that are aggregated at the Census tract level, then please provide them at a different aggregation level, e.g. ZIP code.

[1] https://www2.illinois.gov/rev/localgovernments/Documents/reciprocalagreementoninformationexchangeA-B.pdf

The requested documents will be made available to the general public. This request is being made for commercial purposes.

In the event that there are fees, I would be grateful if you would inform me of the total charges in advance of fulfilling my request. I would prefer the request filled electronically, by e-mail attachment if available or CD-ROM if not.
````

IDOR's response: 

````
Mr. Vance,
 
We do not have data by census tract. Attached is a spreadsheet of sales taxes by Chicago zip codes for 2020, in response to your request.
 
Thank you,
````